// src/app/Models/acountCreation.ts
export interface AccountCreation {
    companyName: string;
    bvnId: string;
    phone: string;
    referralStaff: string;
    cifID: string;
    createdBy: string;
    accountType: string;
    rcNumber: string;
    corporateName: string;
    salutation: string;
    currencyCode: string;
    permanentAddress: string;
    accountOfficerCode: string;
    baselCode: string;
    schmCode: string;
    primarySolId: string;
    glSubHeadCode: string;
    idExpiryDate: string;
    idExpiryDateFieldSpecified: boolean;
    idIssueDate: string;
    emailaddress: string;
    occupation: string;
    preferredBranch: string;
}